requirements.txt
Django>=4.2.0
django-ckeditor>=6.7.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
whitenoise>=6.6.0
python-dotenv>=1.0.0
dj-database-url>=2.1.0
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0
django-admin-interface==0.26.0
django-filebrowser==4.0.1
Pillow==10.0.0
markdown>=3.5.0 